backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
openai
anthropic
pydantic-settings
email-validator
gunicorn
uvicorn[standard]
